changes from feedback
