+++ VideoApi/VideoApi.Events/Handlers/Core/EventHandlerBase.cs
@@ -46,7 +46,7 @@ namespace VideoApi.Events.Handlers.Core
                 .SingleOrDefault(x => x.Id == callbackEvent.ParticipantId);
 
             SourceEndpoint = SourceConference.GetEndpoints().SingleOrDefault(x => x.Id == callbackEvent.ParticipantId);
-            SourceInterpreterRoom = SourceConference.Rooms.SingleOrDefault(x => x is InterpreterRoom && x.Id == callbackEvent.ParticipantRoomId) as InterpreterRoom;
+            SourceInterpreterRoom = SourceConference.Rooms.OfType<InterpreterRoom>().SingleOrDefault(x =>x.Id == callbackEvent.ParticipantRoomId);
             
             await PublishStatusAsync(callbackEvent);
             _logger.LogDebug("Handled callback in {ElapsedMilliseconds}ms", sw.ElapsedMilliseconds);
