+++ VideoApi/VideoApi/appsettings.Development.json
@@ -11,5 +11,5 @@
       }
     }
   },
-  "UseStub": "false"
+  "UseStub": "true"
 }
