+++ VideoApi/VideoApi/appsettings.json
@@ -42,5 +42,5 @@
     "StorageEndpoint": "https://cvpstorageendpoint.net",
     "ManagedIdentityClientId": ""
   },
-  "UseStub": "true"
+  "UseStub": "false"
 }
